Extended deploy test
Part of #227 - extended test for deploy: adapter fails with two sequential deploys now because of incorrect initiation of update process. The test is disabled now and will be enabled as soon as I fix the update. Maven metadata format sent on deploy depends on the package metadata provided by the server.

PR: #228